requirements.txt
appdirs==1.4.3
awscli==1.16.245
boto==2.48.0
botocore==1.12.235
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.2
Django==1.11.5
django-compat==1.0.14
django-cors-middleware==1.3.1
django-geoposition==0.3.0
django-jet==1.0.6
django-openinghours==0.1.3
django-tastypie==0.13.3
django-threadlocals==0.8
docutils==0.15.2
futures==3.3.0
gunicorn==19.7.1
idna==2.8
jmespath==0.9.4
lxml==3.7.3
nose==1.3.7
numpy==1.16.4
packaging==16.8
psycopg2==2.7.1
pyasn1==0.4.7
pycairo==1.18.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2017.2
PyYAML==3.12
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.10.0
urllib3==1.25.6
virtualenv==16.7.5
whitenoise==3.3.0
